Move job sequence props to types module
