+++ apps/prairielearn/src/components/JobSequenceResults.html.tsx
@@ -1,35 +1,11 @@
 import { type HtmlSafeString } from '@prairielearn/html';
 import { hydrateHtml } from '@prairielearn/react/server';
 
-import { ansiToHtml } from '../lib/chalk.js';
 import type { Course } from '../lib/db-types.js';
 import type { JobSequenceWithTokens } from '../lib/server-jobs.types.js';
 
 import { JobSequenceResults } from './JobSequenceResults.js';
-import {
-  type JobSequenceResultsProps,
-  JobSequenceResultsPropsSchema,
-} from './JobSequenceResults.shared.js';
-
-export function getJobSequenceResultsProps({
-  course,
-  jobSequence,
-}: {
-  course: Course | undefined;
-  jobSequence: JobSequenceWithTokens;
-}): JobSequenceResultsProps {
-  return JobSequenceResultsPropsSchema.parse({
-    authnUserUid: jobSequence.authn_user_uid,
-    jobs: jobSequence.jobs.map((job) => ({
-      ...job,
-      outputHtml: ansiToHtml(job.output),
-    })),
-    jobSequence,
-    jobSequenceToken: jobSequence.token,
-    timeZone: course?.display_timezone || 'UTC',
-    userUid: jobSequence.user_uid,
-  });
-}
+import { getJobSequenceResultsProps } from './JobSequenceResults.types.js';
 
 /**
  * Renders the results of a job sequence with live updates.
